changed couble to float

+++ fullyConnectedNN/Accuracyfunctions.h
@@ -6,24 +6,24 @@
 #include <vector>             // For std::vector
 #include <algorithm>   
 
-int argmax(const Eigen::RowVectorXd& vec) {
+int argmax(const Eigen::RowVectorXf& vec) {
     Eigen::Index max_index;
     vec.maxCoeff(&max_index);
     return static_cast<int>(max_index);
 }
 
-double calculateAccuracy(const Eigen::MatrixXd& softmax_output, const Eigen::VectorXd& y_true) {
+float calculateAccuracy(const Eigen::MatrixXf& softmax_output, const Eigen::VectorXf& y_true) {
     int correct = 0;
     for (int i = 0; i < softmax_output.rows(); ++i) {
         int pred = argmax(softmax_output.row(i));
         if (pred == static_cast<int>(y_true(i)))
             ++correct;
     }
-    return static_cast<double>(correct) / softmax_output.rows();
+    return static_cast<float>(correct) / softmax_output.rows();
 }
 
 // Overload for one-hot encoded targets
-double calculateAccuracy(const Eigen::MatrixXd& softmax_output, const Eigen::MatrixXd& y_true) {
+float calculateAccuracy(const Eigen::MatrixXf& softmax_output, const Eigen::MatrixXf& y_true) {
     int correct = 0;
     for (int i = 0; i < softmax_output.rows(); ++i) {
         int pred = argmax(softmax_output.row(i));
@@ -31,9 +31,7 @@ double calculateAccuracy(const Eigen::MatrixXd& softmax_output, const Eigen::Mat
         if (pred == true_class)
             ++correct;
     }
-    return static_cast<double>(correct) / softmax_output.rows();
+    return static_cast<float>(correct) / softmax_output.rows();
 }
-
-
 
 #endif
