+++ fullyConnectedNN/SoftMaxActivation.h
@@ -6,7 +6,7 @@
 class SoftMaxActivation : public ActivationFunction {
     
     public:
-        void forward(const Eigen::MatrixXd& input) override {
+        void forward(const Eigen::MatrixXf& input) override {
             output = (input.array().colwise() - input.rowwise().maxCoeff().array()).exp(); // Stabilize and exponentiate
             output = output.array().colwise() / output.rowwise().sum().array(); // Normalize; // Update input with probabilities
