+++ fullyConnectedNN/SpiralFunction.h
@@ -1,3 +1,5 @@
+// for generating data for testing only 
+
 #ifndef SPIRALFUNCTION_H
 #define SPIRALFUNCTION_H
 
@@ -5,7 +7,7 @@
 
 // Function declaration
 void generate_spiral_data(int points_per_class, int num_classes,
-                          Eigen::MatrixXd& X, Eigen::VectorXd& y,
+                          Eigen::MatrixXf& X, Eigen::VectorXf& y,
                           double noise = 0.2);
 
 #endif
